refactor(core): extract corefuse
