+++ core/commands/mount_windows.go
@@ -17,9 +17,3 @@ var MountCmd = &cmds.Command{
 		res.SetError(errors.New("Mount isn't compatible with Windows yet"), cmds.ErrNormal)
 	},
 }
-
-func Mount(node *core.IpfsNode, fsdir, nsdir string) error {
-	// TODO
-	// currently a no-op, but we don't want to return an error
-	return nil
-}
