+++ core/corefuse/mount_darwin.go
@@ -1,4 +1,4 @@
-package commands
+package corefuse
 
 import (
 	"fmt"
