+++ core/corefuse/mount_windows.go
@@ -1,0 +1,9 @@
+package corefuse
+
+import "github.com/jbenet/go-ipfs/core"
+
+func Mount(node *core.IpfsNode, fsdir, nsdir string) error {
+	// TODO
+	// currently a no-op, but we don't want to return an error
+	return nil
+}
